pyproject.toml
[project]
name = "garmin2strava"
version = "0.1.0"
description = "Garmin to Strava CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click",
    "garminconnect",
    "python-dotenv",
    "requests",
    "stravalib",
    "webbrowser-open>=0.3.0",
    "flask",
    "questionary",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]

[project.scripts]
g2s = "garmin2strava"
